[#190] refactor: 폴더 변경 및 이동
