+++ src/main/java/com/beat/global/external/s3/api/FileApi.java
@@ -1,8 +1,8 @@
-package com.beat.global.external.s3.controller;
+package com.beat.global.external.s3.api;
 
 import com.beat.global.common.dto.ErrorResponse;
 import com.beat.global.common.dto.SuccessResponse;
-import com.beat.global.external.s3.service.dto.PresignedUrlFindAllResponse;
+import com.beat.global.external.s3.application.dto.PresignedUrlFindAllResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
