--- src/main/resources/application-local.yml
@@ -1,21 +1,0 @@
-server:
-  port: 8080
-
-spring:
-  config:
-    activate:
-      on-profile: local
-
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/beat
-    username: root
-    password:
-  jpa:
-    hibernate:
-      ddl-auto: create
-    show-sql: true
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQLDialect
-        format_sql: true
